requirements.txt
Django==1.11
django-bootstrap4==0.0.6
django-rest-framework==0.1.0
djangorestframework==3.8.2
markup==0.2
pytz==2018.4
sinchsms